% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusOne.R
\name{PlusOne}
\alias{PlusOne}
\alias{plusOne}
\title{PlusOne}
\usage{
plusOne(x)
}
\arguments{
\item{x}{A numeric value to be increased by one}
}
\value{
The value of the input plus one
}
\description{
Takes in any numeric value and increases it by one.
}

