% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusOne.R
\name{plusTwo}
\alias{plusTwo}
\title{Plus Two - Long Version}
\usage{
plusTwo(y)
}
\arguments{
\item{y}{A numeric value to be increased by two}
}
\value{
The value of the input plus two
}
\description{
Add two to a number.
}

