% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusOne.R
\name{PlusTwoEasy}
\alias{PlusTwoEasy}
\alias{plusTwoEasy}
\title{Plus Two - Short Version}
\usage{
plusTwoEasy(y)
}
\arguments{
\item{y}{A numeric value to be increased by two}
}
\value{
The value of the numeric argument plus two
}
\description{
Add two to a number.
}

